feat(cache): implement cache check and stream capture hooks for chat functionality
