+++ apps/nebula-chat-server/src/modules/chat/chat.cacheCheck.hook.ts
@@ -1,0 +1,105 @@
+import type { FastifyReply, FastifyRequest, preHandlerAsyncHookHandler } from 'fastify';
+import { cacheService } from '@backend/cache/cache.service';
+import { setCacheHeaders } from '@backend/config/headers.config';
+import { createUserMessage, validateChatRequest } from '@backend/modules/chat/chat.service';
+import type { CreateChatStreamDTO } from '@backend/modules/chat/chat.types';
+import { streamFormatter } from '@backend/modules/chat/chat.utils';
+import { messageService } from '@backend/modules/message/message.service';
+
+export const cacheCheckHook: preHandlerAsyncHookHandler = async (
+  req: FastifyRequest,
+  reply: FastifyReply,
+) => {
+  try {
+    const body = req.body as CreateChatStreamDTO;
+    const key = cacheService.generateKey(body);
+    const cachedData = await cacheService.getFromCache(key);
+
+    if (!cachedData) {
+      return;
+    }
+    // eslint-disable-next-line no-console
+    console.log('Redis: Cache hit');
+
+    const conversationId = body.conversationId;
+    const userMessage = body.messages[0];
+
+    if (!userMessage) {
+      return;
+    }
+
+    await validateChatRequest(conversationId, userMessage);
+
+    const cachedTokens = cachedData.tokens;
+    const lines = cachedTokens.split('\n');
+    let assistantContent = '';
+
+    for (const line of lines) {
+      if (line.startsWith('data: ') && line.includes('"token"')) {
+        try {
+          const dataStr = line.substring(6);
+          const data = JSON.parse(dataStr);
+          if (data.token) {
+            assistantContent += data.token;
+          }
+        } catch {
+          // eslint-disable-next-line no-console
+          console.log('Redis: Failed to parse cached token line');
+        }
+      }
+    }
+
+    const userMessageResult = await createUserMessage(
+      conversationId,
+      userMessage.content,
+      userMessage.role,
+    );
+
+    const assistantMessage = await messageService.createMessage({
+      conversationId: userMessageResult.conversationId,
+      role: 'assistant',
+      content: assistantContent,
+      model: body.model,
+      ...(cachedData.usageData !== undefined && { tokens: cachedData.usageData }),
+    });
+
+    const {
+      userMessageId,
+      conversationId: finalConversationId,
+      isNewConversation,
+    } = userMessageResult;
+    const assistantMessageId = assistantMessage.id;
+
+    reply.hijack();
+    const raw = reply.raw;
+    setCacheHeaders(raw, req.headers.origin);
+    raw.flushHeaders?.();
+
+    if (isNewConversation) {
+      streamFormatter.writeConversationCreated(raw, finalConversationId);
+    }
+
+    streamFormatter.writeUserMessageCreated(raw, userMessageId);
+    streamFormatter.writeCacheHit(raw);
+
+    raw.write(cachedData.tokens.trimEnd());
+    raw.write('\n\n');
+
+    streamFormatter.writeUsage(
+      raw,
+      cachedData.usageData ?? {
+        promptTokens: 0,
+        completionTokens: 0,
+        totalTokens: 0,
+      },
+    );
+
+    streamFormatter.writeAssistantMessageCreated(raw, assistantMessageId);
+    streamFormatter.writeEnd(raw);
+    raw.end();
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.error('Cache check error (fail-open):', error);
+    // Fall through — the handler will run and perform a fresh OpenAI request.
+  }
+};
